move tests to individual files (#1518)
